--- src/main/kotlin/com/comjeonggosi/infra/scheduling/config/SchedulingConfig.kt
@@ -1,8 +1,0 @@
-package com.comjeonggosi.infra.scheduling.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.scheduling.annotation.EnableScheduling
-
-@Configuration
-@EnableScheduling
-class SchedulingConfig
